+++ coverage/collector.py
@@ -10,7 +10,7 @@ from coverage import env
 from coverage.debug import short_stack
 from coverage.disposition import FileDisposition
 from coverage.exceptions import CoverageException
-from coverage.misc import isolate_module
+from coverage.misc import human_sorted, isolate_module
 from coverage.pytracer import PyTracer
 
 os = isolate_module(os)
@@ -352,7 +352,7 @@ class Collector:
             stats = tracer.get_stats()
             if stats:
                 print("\nCoverage.py tracer stats:")
-                for k in sorted(stats.keys()):
+                for k in human_sorted(stats.keys()):
                     print(f"{k:>20}: {stats[k]}")
         if self.threading:
             self.threading.settrace(None)